Fix Emscripten build and TL1 kernel correctness
- Fix CMake ENVIRONMENT quoting that broke Emscripten link step
- Fix WASM SIMD detection bytecode (wrong section sizes caused
  false negative on all browsers)
- Fix benchmark.js load order (define BitNetBench before init())
- Fix TL1 LUT overflow: change from int8 to int16 entries since
  partial sums (a0+a1) can reach 254, exceeding int8 range.
  This eliminates the ~15.7 max error in TL1 kernels.
- Add Docker Compose config for Emscripten builds

+++ src/kernels/gemv.c
@@ -81,7 +81,7 @@ double gemv_run(kernel_type_t kernel,
         case KERNEL_TL1_SIMD: {
             /* Build LUT (included in timing — it's part of the kernel) */
             int32_t lut_size = (K / 2) * 16;
-            int8_t *lut = (int8_t *)calloc(lut_size, sizeof(int8_t));
+            int16_t *lut = (int16_t *)calloc(lut_size, sizeof(int16_t));
             tl1_build_lut(lut, quant_act, K);
 
             if (kernel == KERNEL_TL1_SCALAR) {
